rename Di to De and fix style
