+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 71811,
+  "packagerPid": 4752,
   "expoServerNgrokUrl": "https://ix-2p9.anonymous.nori-nori.exp.direct",
   "packagerNgrokUrl": "https://packager.ix-2p9.anonymous.nori-nori.exp.direct",
-  "ngrokPid": 71828
+  "ngrokPid": 4760
 }
