+++ src/components/MyForm.js
@@ -27,21 +27,22 @@ const styles = StyleSheet.create({
   label: {
     fontWeight: "bold",
     color: colors.white,
-    fontSize: 40
+    fontSize: 32,
+    marginBottom: 8
   },
   input: {
     height: 60,
     width: "90%",
-    fontSize: 30,
-    color: colors.white,
+    fontSize: 32,
+    backgroundColor: colors.white,
+    color: colors.black,
     fontWeight: "bold",
-    borderColor: "gray",
-    borderWidth: 1
+    borderRadius: 8
   },
   error: {
     marginTop: 20,
     marginBottom: 20,
-    fontSize: 20,
+    fontSize: 16,
     fontWeight: "bold",
     color: colors.accent
   }
